refactor: convert start.sh to pure shell and align type names

+++ types/vm.go
@@ -3,6 +3,6 @@ package types
 type VMConfig struct {
 	Name    string `json:"name"`
 	CPU     int    `json:"cpu"`
-	Memory  int    `json:"memory"`
-	Storage int    `json:"storage"`
+	Memory  int64  `json:"memory"`
+	Storage int64  `json:"storage"`
 }
